Move demo app resources to app folder
Organization.
